+++ src/main/java/com/simleetag/homework/api/common/exception/OAuthException.java
@@ -1,0 +1,7 @@
+package com.simleetag.homework.api.common.exception;
+
+public class OAuthException extends RuntimeException {
+    public OAuthException(String message) {
+        super(message);
+    }
+}
