+++ src/main/java/com/simleetag/homework/api/config/JpaConfig.java
@@ -1,0 +1,10 @@
+package com.simleetag.homework.api.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+
+}
