+++ src/main/java/com/simleetag/homework/api/domain/oauth/dto/UserInformationResponse.java
@@ -1,0 +1,17 @@
+package com.simleetag.homework.api.domain.oauth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class UserInformationResponse {
+
+    @JsonProperty("id")
+    private Long id;
+}
