+++ src/main/java/com/simleetag/homework/api/domain/oauth/infra/provider/AbstractOAuthProvider.java
@@ -1,0 +1,15 @@
+package com.simleetag.homework.api.domain.oauth.infra.provider;
+
+import com.simleetag.homework.api.domain.oauth.OAuthAttributes;
+
+abstract class AbstractOAuthProvider implements OAuthProvider {
+    protected OAuthAttributes oauthAttributes;
+
+    protected AbstractOAuthProvider() {
+        this(new OAuthAttributes());
+    }
+
+    protected AbstractOAuthProvider(OAuthAttributes oauthAttributes) {
+        this.oauthAttributes = oauthAttributes;
+    }
+}
