--- src/main/java/com/simleetag/homework/controller/JpaConfig.java
@@ -1,10 +1,0 @@
-package com.simleetag.homework.controller;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-
-}
