--- src/main/java/com/simleetag/homework/domain/oauth/OAuthAttributes.java
@@ -1,14 +1,0 @@
-package com.simleetag.homework.domain.oauth;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class OAuthAttributes {
-    private String clientId;
-    private String clientSecret;
-    private String redirectUri;
-    private String tokenUri;
-    private String userInformationUri;
-}
