--- src/main/java/com/simleetag/homework/dto/UserInformationResponse.java
@@ -1,17 +1,0 @@
-package com.simleetag.homework.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class UserInformationResponse {
-
-    @JsonProperty("id")
-    private Long id;
-}
