--- src/main/java/com/simleetag/homework/exception/OAuthException.java
@@ -1,7 +1,0 @@
-package com.simleetag.homework.exception;
-
-public class OAuthException extends RuntimeException {
-    public OAuthException(String message) {
-        super(message);
-    }
-}
