--- src/test/java/com/simleetag/homework/controller/mockmvc/WebMockMvc.java
@@ -1,11 +1,0 @@
-package com.simleetag.homework.controller.mockmvc;
-
-import org.springframework.test.web.servlet.RequestBuilder;
-import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.request.RequestPostProcessor;
-
-public interface WebMockMvc {
-    ResultActions perform(RequestBuilder requestBuilder) throws Exception;
-
-    RequestPostProcessor userToken();
-}
